[CHORE] FE: web 구조 변경
1. fragment 분리

2. 메뉴 클릭 이벤트 header.js 로 분리
