+++ src/main/webapp/WEB-INF/views/common/layout/fragments/footer-new.html
@@ -1,0 +1,39 @@
+<footer th:fragmen="footer" id="footer">
+    <div class="footerTopWrap">
+        <div class="footerTop">
+            <img src="/static/images/common/logo_full_w.png">
+            <div class="keywordBox">
+                <p>
+                    MY WORLD,
+                </p>
+                <p>
+                    FRIENDSHIP,
+                </p>
+                <p>
+                    MEMORIES.
+                </p>
+            </div>
+        </div>
+    </div>
+    <div class="footerBotWrap">
+        <div class="footerBottom">
+            <div class="footerNav">
+                <ul>
+                    <li id="miniHompi2" style="cursor: pointer;">내미니홈피</li>
+                    <li id="giftPage2" style="cursor: pointer;">선물가게</li>
+                    <li id="cartPage2" style="cursor: pointer;">장바구니</li>
+                    <li id="noticePage2" style="cursor: pointer;">공지사항</li>
+                    <li id="aboutUsPage2" style="cursor: pointer;">ABOUT US</li>
+                </ul>
+            </div>
+            <div class="copyright">
+                <p>
+                    <span>ⓒ2025.Team.</span>
+                    <span class="teamName">WannabeDev</span>
+                    <span>All rights reserved.</span>
+                </p>
+                <p>Since 2025.02.24</p>
+            </div>
+        </div>
+    </div>
+</footer>
