+++ src/main/webapp/WEB-INF/views/common/layout/fragments/head-new.html
@@ -1,0 +1,26 @@
+<head th:fragment="head">
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <meta http-equiv="X-UA-Compatible" content="ie=edge">
+    <meta name="description" content="워너비">
+    <title>워너비</title>
+    
+    <!-- library -->
+    
+    <link href="http://cdn.jsdelivr.net/xeicon/2/xeicon.min.css" rel="stylesheet">
+    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/swiper@11/swiper-bundle.min.css"/>
+    <link rel="preconnect" href="https://fonts.googleapis.com">
+    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
+    <link href="https://fonts.googleapis.com/css2?family=Jua&family=Nanum+Gothic&display=swap" rel="stylesheet">
+    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/swiper@11/swiper-bundle.min.css" />
+    
+    <!-- custom common  -->
+    
+    <link rel="stylesheet" th:href="@{/static/css/common/reset.css}"/>
+    <link rel="stylesheet" th:href="@{/static/css/common/onlineStyle.css}"/>
+    <link rel="stylesheet" th:href="@{/static/css/common/common.css}"/>
+    <link rel="stylesheet" th:href="@{/static/css/common/header.css}"/>
+    <link rel="stylesheet" th:href="@{/static/css/common/footer.css}"/>
+
+</head>
+
