+++ src/main/webapp/WEB-INF/views/common/layout/fragments/header-new.html
@@ -1,0 +1,39 @@
+<header th:fragment="header" class="header">
+    <div class="topBar">
+        <div class="topBarLogo">
+            <a href="/" class="logo"><img src="/static/images/common/logo_full.png"></a>
+        </div>
+        <div class="topBarInfo ">
+            <div class="topBarList" th:if="${session.userData != null}">
+                <ul>
+                    <li id="loginIp" class="ipDisplay" th:text="${session.userData.accessIp}">아이피</li>
+                    <li id="loginName" class="nameDisplay" th:text="${session.userData.name}">이름</li>
+                    <li><a href="/"><img src="/static/images/common/homeIcon.png"></a></li>
+                    <li><a href="/userUpdate"><img src="/static/images/common/option.png"></a></li>
+                    <li><a href="#" id="logoutBtn" class="btn nobtn logoutBtn">로그아웃</a></li>
+                </ul>
+            </div>
+            <div class="topBarList" th:if="${session.userData == null}">
+                <ul>
+                    <li><input class="loginIdInput" type="text" id="loginId" placeholder="아이디"></li>
+                    <li><input class="passwordInput" type="password" id="password" placeholder="비밀번호"></li>
+                    <li><a href="#" id="loginBtn" class="btn loginBtn">로그인</a></li>
+                    <li><a href="#" id="signupBtn" class="signupBtn btn">회원가입</a></li>
+                </ul>
+            </div>
+        </div>
+    </div>
+    <!--헤더 끝-->
+    
+    <!--gnb 시작-->
+    <div class="gnb">
+        <menu class="menuWrap">
+            <li id="miniHompi">내미니홈피</li>
+            <li id="giftPage">선물가게</li>
+            <li id="cartPage">장바구니</li>
+            <li id="noticePage">공지사항</li>
+            <li id="aboutUsPage">ABOUT US</li>
+        </menu>
+    </div>
+    <!--gnb 끝-->
+</header>
