+++ src/main/webapp/WEB-INF/views/common/layout/fragments/script-new.html
@@ -1,0 +1,18 @@
+
+<div th:fragment="script">
+    
+    <!-- library -->
+    <script src="https://code.jquery.com/jquery-3.7.1.min.js"
+            integrity="sha256-/JqT3SQfawRcv/BIHPThkBvs0OEvtFFmqPF/lYI/Cxo=" crossorigin="anonymous"></script>
+    <script src="https://cdn.jsdelivr.net/npm/sweetalert2@11"></script>
+    
+    <!-- custom common -->
+    <script th:type="module" th:src="@{/static/js/ptr/api.js}"></script>
+    <script th:type="module" th:src="@{/static/js/ptr/page.js}"></script>
+    <script th:type="module" th:src="@{/static/js/ptr/httpStatus.js}"></script>
+    <script th:type="module" th:src="@{/static/js/ptr/swal.js}"></script>
+    <script th:type="module" th:src="@{/static/js/ptr/ajax.js}"></script>
+    <script th:type="module" th:src="@{/static/js/ptr/header.js}"></script>
+    <script th:type="module" th:src="@{/static/js/ptr/login-new.js}"></script>
+
+</div>
