+++ src/main/webapp/WEB-INF/views/common/layout/layout.html
@@ -35,7 +35,7 @@
 </div>
 <!-- 공통 JS -->
 <script src="/static/js/common.js"></script>
-<script src="/static/js/home/login.js"></script>
+<script src="/static/js/home/loginNew.js"></script>
 <!--<script src="/static/js/home/moveFromHome.js"></script>--> <!-- 파일 없는데? 안쓰는거면 지워주세요 -->
 <!-- 개별 페이지 JS -->
 <th:block layout:fragment="scripts"></th:block>
