+++ src/main/webapp/WEB-INF/views/home/notice/noticeView.html
@@ -7,7 +7,7 @@
     <link href="http://cdn.jsdelivr.net/xeicon/2/xeicon.min.css" rel="stylesheet">
     <script src="https://code.jquery.com/jquery-3.7.1.min.js"
             integrity="sha256-/JqT3SQfawRcv/BIHPThkBvs0OEvtFFmqPF/lYI/Cxo=" crossorigin="anonymous"></script>
-    <script src="/static/js/login.js"></script>
+    <script src="/static/js/loginNew.js"></script>
 
     <link rel="stylesheet" href="/static/css/notice/noticeView.css">
     <th:block th:replace="common/layout/fragments/favicon :: favicon"></th:block>
