+++ src/main/webapp/WEB-INF/views/sample/onlineMain_yesLogin.html
@@ -45,7 +45,7 @@
 		<div id="container">
 			<div class="containerWrap">
 				<div class="mainTop">
-					<!--login s-->
+					<!--loginNew s-->
 					<div class="mainTopLeft">
 						<div class="loginWrap yesLogin">
 							<div class="yesLoginDiv1">
@@ -94,7 +94,7 @@
 							<a class="btn">내 미니홈피 바로가기</a>
 						</div>
 					</div>
-					<!--login e-->
+					<!--loginNew e-->
 					<div class="mainTopCenter">
 						메인슬라이더
 					</div>
@@ -298,7 +298,7 @@
 		</div>
 		<!--푸터 끝-->
 	</div>
-	<script src="/js/home/login.js"></script>
+	<script src="/js/home/loginNew.js"></script>
  </body>
 </html>
