+++ src/main/webapp/static/js/ptr/main-new.js
@@ -1,21 +1,9 @@
 // main home.js
 
 import API from './api.js';
-import VIEW from "./page.js";
-import { getAjax, postAjax, putAjax, deleteAjax, fetchThen } from "./ajax.js";
+import { getAjax, fetchThen } from "./ajax.js";
 import { swalPopup } from "./swal.js";
 import Swiper from 'https://cdn.jsdelivr.net/npm/swiper@11/swiper-bundle.min.mjs'
-
-// =========== 변수 레이어 ===========
-
-// 메뉴 관련 요소
-const miniHompi             = $("#miniHompi,    #miniHompi2");
-const giftPage              = $("#giftPage,     #giftPage2");
-const cartPage              = $("#cartPage,     #cartPage2");
-const noticePage            = $("#noticePage,   #noticePage2");
-const aboutUsPage           = $("#aboutUsPage,  #aboutUsPage2");
-const signUpButtons         = $("#signupBtn,    #signupBtn2");
-const dotoriRechargeBtn     = $("#rechargePageBtn");
 
 
 // 미니홈피 팝업 스펙
@@ -66,8 +54,6 @@ const movieTemplate = `
 
 // =========== 실행 레이어 ===========
 $(document).ready(() => {
-    // 메뉴 클릭 이벤트 설정
-    componentClickEvent();
 
     // 슬라이드 배너 렌더링
     renderSlideBanner();
@@ -97,17 +83,6 @@ $(document).ready(() => {
 });
 
 // =========== 함수 레이어 ===========
-
-// 메뉴 클릭 이벤트: 각 요소 클릭 시 지정된 페이지로 이동하거나 기능 실행
-function componentClickEvent() {
-    miniHompi.click(()          => openPop());
-    giftPage.click(()           => (location.href = VIEW.GIFT_SHOP));
-    cartPage.click(()           => (location.href = VIEW.CART));
-    noticePage.click(()         => (location.href = VIEW.NOTICE));
-    aboutUsPage.click(()        => (location.href = VIEW.ABOUT_US));
-    signUpButtons.click(()      => (location.href = VIEW.SIGN_UP));
-    dotoriRechargeBtn.click(()  => (location.href = VIEW.RECHARGE));
-}
 
 // 미니홈피 팝업 오픈 함수
 function openPop() {
@@ -151,7 +126,8 @@ function renderSlideBanner() {
             $('.swiper .swiper-wrapper').html(slidesHtml);
 
             new Swiper('.swiper', {
-                spaceBetween: 30,
+                // spaceBetween: 30,
+                effect : "fade",
                 centeredSlides: true,
                 autoplay: {
                     delay: 2500,
@@ -161,10 +137,10 @@ function renderSlideBanner() {
                     el: '.swiper-pagination',
                     clickable: true,
                 },
-                navigation: {
+               /* navigation: {
                     nextEl: '.swiper-button-next',
                     prevEl: '.swiper-button-prev',
-                },
+                },*/
             });
         },
         (error) => {
